server with events 2.1.1, fix: camera length was accidentally polled as an int
* in fact the `indicated_camera_length` is a float in unit meters
* e.g., 0.028999 for camera length `"29 mm"` in diffraction mode

+++ scripts/start-remote-dummy-server-with-events-legacy-titan.py
@@ -41,7 +41,7 @@ try:
         # for meta data key 'electron_gun.voltage'
         "voltage": (float, 1),                  # e.g., "200"
         # for meta data key "objective.mode -> projector.camera_length"
-        "indicated_camera_length": (int, 1),    # e.g., "0", in meters (?)
+        "indicated_camera_length": (float, 1),  # e.g., "0.028999", in meters
         # for meta data key "objective.mode -> projector.magnification"
         "indicated_magnification": (float, 1),  # e.g., 200000.0
         # for meta data key "objective.mode -> projector.mode"
